changed resume summary to entire content for cover letter
