+++ generate_job_search.py
@@ -6,6 +6,7 @@ from langchain.embeddings import OpenAIEmbeddings
 from langchain.document_loaders import CSVLoader
 from langchain.vectorstores import DocArrayInMemorySearch
 from langchain.chains import RetrievalQA
+from langchain_utils import get_index
 
 
 from dotenv import load_dotenv, find_dotenv
@@ -50,6 +51,25 @@ def find_similar_jobs(job_title):
 #TBD: based on resume information, suggest jobs 
 def generate_job_suggestion():
     return None
+
+
+def extract_resume_fields(resume_file):
+
+    index = get_index(resume_file)
+
+    query = f"""" Search and extract fields of this resume. 
+
+    Some common resume fields include but not limited to personal information, objective, education, work experience, awards and honors, and skills.
+
+    List all the field information in a markdown table.
+    
+    """
+
+    response = index.query(query)
+    print(response)
+    return response
+
+
 
 my_job_title = 'software developer'
 if __name__ == '__main__':
